Create ModelBase class
