+++ src/pdag/_node.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ._model import ParameterBase, Relationship
+    from ._base import ParameterBase, Relationship
 
 
 class ParameterNode[T]:
